+++ lib/src/components/input_field/input_field_style.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/widgets.dart';
 import 'package:impaktfull_ui/src/models/asset.dart';
 import 'package:impaktfull_ui/src/theme/theme.dart';
+import 'package:impaktfull_ui/src/util/extension/color_extensions.dart';
 import 'package:impaktfull_ui/src/util/extension/text_style_extension.dart';
 
 class ImpaktfullUiInputFieldTheme extends ImpaktfullUiComponentTheme {
@@ -50,7 +51,7 @@ class ImpaktfullUiInputFieldTheme extends ImpaktfullUiComponentTheme {
           border: colors.border,
           borderError: colors.error,
           cursor: colors.accent,
-          selection: colors.accent.withOpacity(0.3),
+          selection: colors.accent.withOpacityPercentage(0.3),
           selectionHandle: colors.accent,
         ),
         dimens: ImpaktfullUiInputFieldDimensTheme(
@@ -61,7 +62,7 @@ class ImpaktfullUiInputFieldTheme extends ImpaktfullUiComponentTheme {
           text: textStyles.onCard.text.medium,
           placeholder: textStyles.onCardTertiary.text.medium.copyWith(
               color: textStyles.onCardTertiary.text.medium.color
-                  ?.withOpacity(0.5)),
+                  ?.withOpacityPercentage(0.5)),
           hint: textStyles.onCardTertiary.text.small,
           error: textStyles.onCardDestructive.text.small.medium
               .copyWith(color: colors.error),