+++ lib/src/components/modal/modal_style.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/widgets.dart';
-import 'package:impaktfull_ui/impaktfull_ui.dart';
 import 'package:impaktfull_ui/src/models/asset.dart';
 import 'package:impaktfull_ui/src/theme/theme.dart';
+import 'package:impaktfull_ui/src/util/extension/text_style_extension.dart';
 
 class ImpaktfullUiModalTheme extends ImpaktfullUiComponentTheme {
   final ImpaktfullUiModalColorTheme colors;