+++ lib/src/util/extension/color_extensions.dart
@@ -1,0 +1,6 @@
+import 'dart:ui';
+
+extension ColorExtension on Color {
+  Color withOpacityPercentage(double percentage) =>
+      withAlpha((255.0 * percentage).round());
+}